Refactor utils into smaller, independent files
