+++ superset/assets/src/utils/common.js
@@ -1,8 +1,5 @@
-/* eslint global-require: 0 */
+import d3 from 'd3';
 import { SupersetClient } from '@superset-ui/core';
-import { t } from '@superset-ui/translation';
-
-const d3 = require('d3');
 
 export const EARTH_CIRCUMFERENCE_KM = 40075.16;
 export const LUMINANCE_RED_WEIGHT = 0.2126;
@@ -112,11 +109,6 @@ export function optionFromValue(opt) {
   // From a list of options, handles special values & labels
   return { value: optionValue(opt), label: optionLabel(opt) };
 }
-
-// Error messages used in many places across applications
-export const COMMON_ERR_MESSAGES = {
-  SESSION_TIMED_OUT: t('Your session timed out, please refresh your page and try again.'),
-};
 
 // time_range separator
 export const TIME_RANGE_SEPARATOR = ' : ';
