+++ superset/assets/src/utils/errorMessages.js
@@ -1,0 +1,6 @@
+// Error messages used in many places across applications
+const COMMON_ERR_MESSAGES = {
+  SESSION_TIMED_OUT: 'Your session timed out, please refresh your page and try again.',
+};
+
+export default COMMON_ERR_MESSAGES;
